% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kmean}
\alias{kmean}
\title{K Means a Dataset}
\usage{
kmean(Data)
}
\arguments{
\item{Data}{A data table to be clustered}
}
\value{
The data with the fit cluster field
}
\description{
Takes in any data table and does a k means cluster to the data
}

